# docs/04-api/openapi.yaml
openapi: 3.0.3
info:
  title: Employee Service API
  version: 1.1.0
  description: |
    **Frozen Target API contract** — source of truth is the current runtime
    (`app/api`, `app/schemas`, `app/services`) and tests. Sibling docs:
    `api-contract-analysis.md`, `approval-api-contract.md`.

    **Target:** JWT Bearer (ADR-AUTH-JWT-01), integer business IDs + UUID User
    (ADR-ID-01), nested error envelope (ADR-ERR-03), live config without
    snapshot fields (ADR-LIVE-CFG-01), Action Engine via
    `GET .../available-actions` + `POST .../actions/{action_id}` (ADR-ACTION-01).

    Approver flow: `GET /requests/{id}` → `GET /requests/{id}/available-actions`
    → `POST /requests/{id}/actions/{action_id}`.

    `/approval-tasks/*` routes are **disabled compatibility stubs** (`409 INVALID_STATE`).
    Legacy `POST .../submit` and `POST .../cancel` are **deprecated thin aliases**
    of Action Engine (response = `RequestCard`).
...
security:
  - bearerAuth: []
...
paths:
...
  /request-types:
    get:
      tags: [RequestTypes]
      summary: Get active request types
      description: |
        Returns only active request types. Empty catalog → `200` and `[]`.
        Roles: `employee`, `admin`. Traceability: UC-03; FR-CAT-01.
      responses:
        "200":
          description: Active request types.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestTypeSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /request-types/{type_id}:
    get:
      tags: [RequestTypes]
      summary: Get an active request type
      description: |
        Absent or inactive type → `404 NOT_FOUND`.
        Roles: `employee`, `admin`. Traceability: UC-03; FR-CAT-02.
      parameters:
        - $ref: "#/components/parameters/TypeId"
      responses:
        "200":
          description: Active request type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestTypeSummary"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`NOT_FOUND` — type does not exist or is inactive."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /request-types/{type_id}/schema:
    get:
      tags: [RequestTypes]
      summary: Get the live request form schema
      description: |
        Live form schema for an active type. Absent/inactive → `404 NOT_FOUND`.
        Roles: `employee`, `admin`. Traceability: UC-03, UC-04; FR-CAT-03.
      parameters:
        - $ref: "#/components/parameters/TypeId"
      responses:
        "200":
          description: Current live form schema.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestTypeSchema"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`NOT_FOUND` — type does not exist or is inactive."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests:
    post:
      tags: [Requests]
      summary: Create a draft request
      description: |
        Creates a draft for an active request type. Initiator = current user.
        Creates a HistoryEvent (`create` → draft) atomically.
        Role: `employee`. Traceability: UC-04; FR-REQ-01.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestInput"
            example:
              request_type_id: 1
      responses:
        "201":
          description: Draft request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`NOT_FOUND` — request type does not exist."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: "`INACTIVE_TYPE` — request type is inactive."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: "`VALIDATION` — invalid request_type_id."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags: [Requests]
      summary: List the current employee's requests
      description: |
        Own requests only. Sorted by `created_at` descending.
        Role: `employee`. Traceability: UC-06; FR-CAB-02.
      parameters:
        - $ref: "#/components/parameters/RequestStatusFilter"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Request list items (array; pagination envelope not fixed).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestListItem"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: "`VALIDATION` — unknown status or invalid page/page_size."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /requests/{request_id}:
    get:
      tags: [Requests]
      summary: Get a visible request card
      description: |
        Returns `RequestCard` when visible under ACL:
        - `employee` — own request;
        - `approver` — request with any own ApprovalTask (any task status).
        Invisible or missing request → `404 NOT_FOUND`.
        Role `admin` is not permitted on this endpoint (role gate).
        `available_actions` are **not** embedded; use
        `GET /requests/{request_id}/available-actions`.
        Form schema is **not** embedded; use `/request-types/{type_id}/schema`.
        Traceability: UC-06, UC-07; FR-REQ-04; BR-14.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Request card.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestCard"
              example:
                id: 10245
                request_type:
                  id: 1
                  code: annual_leave
                  name: Annual leave
                initiator_user_id: 33333333-3333-4333-8333-333333333333
                initiator:
                  id: 33333333-3333-4333-8333-333333333333
                  full_name: Иван Петров
                status_id: 2
                status:
                  id: 2
                  code: in_approval
                  name: На согласовании
                current_stage_id: 10
                current_stage:
                  id: 10
                  name: Line manager
                  sequence_no: 1
                created_at: "2026-09-21T09:00:00Z"
                updated_at: "2026-09-21T09:10:00Z"
                values: []
                approval_tasks:
                  - id: 501
                    stage_id: 10
                    assignee_user_id: 77777777-7777-4777-8777-777777777777
                    status: open
                    created_at: "2026-09-21T09:10:00Z"
                    completed_at: null
                comments: []
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RequestNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Requests]
      summary: Save working values (deferred stub)
      description: |
        **Deferred stub in current runtime.** Always returns `409 INVALID_STATE`
        («сохранение полей будет доступно после следующего этапа API»).
        Not Target-active. Role gate: `employee`.
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateValuesInput"
      responses:
        "409":
          description: "`INVALID_STATE` — deferred stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /approval-tasks:
    get:
      tags: [ApprovalTasks]
      summary: "[Disabled stub] List approval tasks"
      description: |
        **Disabled compatibility stub.** Always returns `409 INVALID_STATE`.
        Not Target-active. Not a thin alias.
        Approver queue is not implemented; use request card ACL + available-actions.
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "409":
          description: "`INVALID_STATE` — disabled stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: INVALID_STATE
                  message: Очередь согласования будет доступна после следующего этапа API
                  details: {}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /approval-tasks/{task_id}:
    get:
      tags: [ApprovalTasks]
      summary: "[Disabled stub] Get approval task card"
      description: |
        **Disabled compatibility stub.** Always returns `409 INVALID_STATE`.
        Not Target-active. Approver reads `GET /requests/{request_id}` instead.
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "409":
          description: "`INVALID_STATE` — disabled stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /approval-tasks/{task_id}/approve:
    post:
      tags: [ApprovalTasks]
      summary: "[Disabled stub] Approve via task path"
      description: |
        **Disabled compatibility stub.** Always returns `409 INVALID_STATE`.
        Not a thin alias. Use `POST /requests/{id}/actions/{action_id}`.
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteActionInput"
      responses:
        "409":
          description: "`INVALID_STATE` — disabled stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /approval-tasks/{task_id}/reject:
    post:
      tags: [ApprovalTasks]
      summary: "[Disabled stub] Reject via task path"
      description: |
        **Disabled compatibility stub.** Always returns `409 INVALID_STATE`.
        Not a thin alias. Use `POST /requests/{id}/actions/{action_id}`.
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteActionInput"
      responses:
        "409":
          description: "`INVALID_STATE` — disabled stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /approval-tasks/{task_id}/return:
    post:
      tags: [ApprovalTasks]
      summary: "[Disabled stub] Return via task path"
      description: |
        **Disabled compatibility stub.** Always returns `409 INVALID_STATE`.
        Not a thin alias. Use `POST /requests/{id}/actions/{action_id}`.
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteActionInput"
      responses:
        "409":
          description: "`INVALID_STATE` — disabled stub."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Target authentication (ADR-AUTH-JWT-01).
  parameters:
    TypeId:
      name: type_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Request type ID (integer).
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Request ID (integer; also the UI request number).
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: ApprovalTask ID (integer). Stub routes only.
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    RequestStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/RequestStatus"
  responses:
    Unauthorized:
      description: "`UNAUTHORIZED` — missing or invalid JWT."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Требуется аутентификация
              details: {}
    Forbidden:
      description: "`FORBIDDEN` — role does not permit the operation."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Недостаточно прав для выполнения операции
              details: {}
...
    RequestNotFound:
      description: "`NOT_FOUND` — request does not exist or is not visible to the caller."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: Объект не найден
              details: {}
    InternalError:
      description: "`INTERNAL` — unexpected server error; mutations are rolled back."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL
              message: Произошла внутренняя ошибка. Попробуйте позже
              details: {}
  schemas:
    RequestStatus:
      type: string
      enum: [draft, in_approval, returned, approved, rejected, cancelled]
...
    ErrorCode:
      type: string
      description: Target error codes without ERR_ prefix (ADR-ERR-03).
      enum:
        - UNAUTHORIZED
        - INVALID_CREDENTIALS
        - FORBIDDEN
        - FORBIDDEN_APPROVAL
        - NOT_FOUND
        - VALIDATION
        - INACTIVE_TYPE
        - INVALID_STATE
        - REQUEST_ACTION_NOT_ALLOWED
        - TASK_DONE
        - ROUTE_CONFIG
        - INTERNAL
    ErrorBody:
      type: object
      required: [code, message, details]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      description: Nested error envelope (ADR-ERR-03).
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    StatusRef:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
...
    RequestTypeSummary:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
    RequestTypeRef:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
    RequestFieldDefinition:
      type: object
      required: [code, name, data_type, required, order_no, dictionary_id]
      properties:
        code:
          type: string
        name:
          type: string
        data_type:
          type: string
          enum: [text, date, number, catalog, boolean]
        required:
          type: boolean
        order_no:
          type: integer
        dictionary_id:
          type: integer
          format: int64
          nullable: true
    RequestTypeSchema:
      type: object
      required: [request_type_id, fields]
      properties:
        request_type_id:
          type: integer
          format: int64
        fields:
          type: array
          items:
            $ref: "#/components/schemas/RequestFieldDefinition"
    CurrentStage:
      type: object
      required: [id, name, sequence_no]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        sequence_no:
          type: integer
    UserReference:
      type: object
      required: [id, full_name]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
    FieldValue:
      type: object
      required: [field_code, value]
      properties:
        field_code:
          type: string
        value:
          type: string
          nullable: true
    ApprovalTaskSummary:
      type: object
      required: [id, stage_id, assignee_user_id, status, created_at]
      properties:
        id:
          type: integer
          format: int64
        stage_id:
          type: integer
          format: int64
        assignee_user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [open, completed, cancelled]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    Comment:
      type: object
      required: [id, kind, text, author, created_at]
      properties:
        id:
          type: integer
          format: int64
        kind:
          type: string
          enum: [free, decision]
        text:
          type: string
        author:
          $ref: "#/components/schemas/UserReference"
        approval_task_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateRequestInput:
      type: object
      required: [request_type_id]
      additionalProperties: false
      properties:
        request_type_id:
          type: integer
          format: int64
    CreatedRequest:
      type: object
      required:
        - id
        - request_type_id
        - initiator_user_id
        - status_id
        - status
        - current_stage_id
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        request_type_id:
          type: integer
          format: int64
        initiator_user_id:
          type: string
          format: uuid
        status_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RequestListItem:
      type: object
      required: [id, request_type, status, current_stage, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        request_type:
          $ref: "#/components/schemas/RequestTypeRef"
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage:
          allOf:
            - $ref: "#/components/schemas/CurrentStage"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RequestCard:
      type: object
      description: |
        Target request card (runtime). No embedded schema, available_actions,
        value_source, submit_number, or field_value_versions.
      required:
        - id
        - request_type
        - initiator_user_id
        - initiator
        - status_id
        - status
        - current_stage_id
        - created_at
        - updated_at
        - values
        - approval_tasks
        - comments
      properties:
        id:
          type: integer
          format: int64
        request_type:
          $ref: "#/components/schemas/RequestTypeRef"
        initiator_user_id:
          type: string
          format: uuid
        initiator:
          $ref: "#/components/schemas/UserReference"
        status_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/StatusRef"
        current_stage_id:
          type: integer
          format: int64
          nullable: true
        current_stage:
          allOf:
            - $ref: "#/components/schemas/CurrentStage"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        values:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
        approval_tasks:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalTaskSummary"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    UpdateValuesInput:
      type: object
      required: [values]
      additionalProperties: false
      properties:
        values:
          type: array
          items:
            $ref: "#/components/schemas/FieldValue"
    ExecuteActionInput:
      type: object
      additionalProperties: false
      properties:
        comment:
          type: string
          nullable: true
          description: |
            Required (non-empty after trim) for reject/return; optional otherwise.
...